Support for tags in AstGen. Fixes issue 7.
Nonnull is now implemented using a tag. The syntax sugar "!" is
still supported, but is equivalent to prepending the member by
"[nonnull]". A member can be tagged with multiple tags by saying
"[tagA] [tagB]".

An example of how tags are used is in the test tags.list. Note
that in tag expressions | and & have the SAME precedence and
left-associativity.