[branch] BUG: AstGen crashes with unbalanced {}
New issue
Summary: crash when {} are not balanced
Labels: Type-Defect Priority-Low OpSys-All Package-AstGen Milestone-Release0.3
Status: New
If there are more } than { the TemplateParser tries to rewind
when there is no mark. I believe this used to be a no-op, but
the proper fix would be to fix the TemplateParser to not try to
rewind the token stream when it didn't set a mark on it.