Better location tracking.
Update issue 23
You can now log the places from which AST nodes are built
without exlicitely giving a location by saying:
  fb file.bpl -lf=log -lc=ast
A few places in the parser where this was done were fixed.
However, I now believe that it is better to always say
  FileLocation.unknown()
explicitely if the location is not known. This is not high
priority, so I just leave this issue open.