TypeChecker now uses the memo cache in Evaluator.
The method types() is a bit safer because it makes a copy
of the cache (so clients can't arbitrarily manipulate 
typechecker's cache). However, since the type is slightly
different a copy *must* be made. If this is a performance
problem then I'll need, perhaps, to parametrize Evaluator
by the type of keys used in the memo cache.